IR: move BB predecessors computation as a seperate analysis pass
Basic block predecessor sets are easy to compute when building a function from
scratch, but they are tricky to update when optimization passes affects the
CFG. Besides CFG changes are not atomic, so at some point the CFG may be
invalid and it's not clear when to update predecessors.

Move predecessors computation as a separate pass so that all predecessors
consumers compute this information as needed and have to cope themselves with
data invalidation.